// backend/src/modules/product/events/product.event.types.ts
export enum ProductEventType {
  CREATED = 'product.created',
  UPDATED = 'product.updated',
  STATUS_CHANGED = 'product.status_changed',
  DELETED = 'product.deleted',
  VIEWED = 'product.viewed',
  IMAGE_ADDED = 'product.image_added',
  VARIANT_ADDED = 'product.variant_added',
  VARIANT_UPDATED = 'product.variant_updated',
  VARIANT_DELETED = 'product.variant_deleted',
  ATTRIBUTES_SYNCED = 'product.attributes_synced',
}